+++ ArcdpsLogManager/Logs/Updates/LogDataUpdater.cs
@@ -43,7 +43,12 @@ namespace GW2Scratch.ArcdpsLogManager.Logs.Updates
 				"Support for Ai, Keeper of the Peak"),
 			new LogUpdate(log => log.ParsingVersion < new Version(1, 0, 3, 1)
 					&& log.Encounter == Encounter.KeepConstruct,
-				"Keep Construct CM detection"),
+				"Adds Keep Construct CM detection."),
+			new LogUpdate(log => log.ParsingVersion < new Version(1, 0, 3, 2)
+					&& log.Encounter == Encounter.Xera
+					&& log.EncounterResult == EncounterResult.Failure
+					&& log.HealthPercentage > 0.998,
+				"Twisted Castle logs may have been incorrectly identified as Xera."),
 			// When adding a new update, you need to increase the revision (last value) of the version in the .csproj file
 			// unless the version changes more significantly, in that case it can be reset to 0.
 		};